// api/generated/model/measurementFormDtoUnit.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * LgymApi.Api
 * OpenAPI spec version: 1.0
 */

export type MeasurementFormDtoUnit = typeof MeasurementFormDtoUnit[keyof typeof MeasurementFormDtoUnit];


export const MeasurementFormDtoUnit = {
  Unknown: 'Unknown',
  Kilograms: 'Kilograms',
  Pounds: 'Pounds',
  Meters: 'Meters',
  Centimeters: 'Centimeters',
  Millimeters: 'Millimeters',
  Percent: 'Percent',
  Bmi: 'Bmi',
} as const;