// api/generated/model/measurementResponseDto.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * LgymApi.Api
 * OpenAPI spec version: 1.0
 */
import type { EnumLookupDto } from './enumLookupDto';

export interface MeasurementResponseDto {
  /** @nullable */
  user?: string | null;
  bodyPart?: EnumLookupDto;
  unit?: EnumLookupDto;
  value?: number;
  /** @nullable */
  createdAt?: string | null;
  /** @nullable */
  updatedAt?: string | null;
}