// api/generated/model/measurementTrendDto.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * LgymApi.Api
 * OpenAPI spec version: 1.0
 */
import type { EnumLookupDto } from './enumLookupDto';

export interface MeasurementTrendDto {
  bodyPart?: EnumLookupDto;
  unit?: EnumLookupDto;
  /** @nullable */
  firstMeasurementValue?: number | null;
  /** @nullable */
  firstMeasurementDate?: string | null;
  /** @nullable */
  lastMeasurementValue?: number | null;
  /** @nullable */
  lastMeasurementDate?: string | null;
  /** @nullable */
  difference?: number | null;
  /** @nullable */
  startValue?: number | null;
  /** @nullable */
  currentValue?: number | null;
  /** @nullable */
  change?: number | null;
  /** @nullable */
  changePercentage?: number | null;
  /** @nullable */
  direction?: string | null;
  points?: number;
}